+++ src/app/components/Comparison.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import FetchData from '../logics/FetchData';
-import GenerateStats from '../logics/GenerateStats';
+import FetchUserDataPromise from '../logics/FetchUserDataPromise';
 
 class Comparison extends React.Component {
   constructor() {
@@ -9,26 +8,9 @@ class Comparison extends React.Component {
   }
 
   process() {
-    const fetchPromise = (username) => {
-      FetchData(username, true)
-        .then((response) => {
-          const data = JSON.parse(response);
-          if (data.success) {
-            return data;
-          }
-          throw new Error('User not found');
-        })
-        .then(data => GenerateStats(data))
-        .then((data) => {
-          console.log(data);
-        })
-        .catch((err) => {
-          console.log(err);
-        });
-    };
     Promise.all([
-      fetchPromise(this.props.match.params.username1),
-      fetchPromise(this.props.match.params.username2),
+      FetchUserDataPromise(this.props.match.params.username1),
+      FetchUserDataPromise(this.props.match.params.username2),
     ])
       .then((responses) => {
         console.log(responses[0]);
